Patch fix for kernel 4.9
While working with kernel 4.9, I face with error and searched about it. The error was about "too many arguments". For kernel 4.12 and 4.9 the function cfg80211_roamed takes 3 arguments rather than 8 which is for higher kernel versions.
While searching for fix I found forums.kali.org post which had fix for this issue.
Reference: https://forums.kali.org/showthread.php?37911-Getting-RTL8188-to-work-with-Kali-in-monitor-mode

The same issue exists for rtl8188eus driver for kernel 4.9 and 4.12 as well but since this driver wasn't implemented in patches, I didn't add it.